SSCSCI-9: go back to original secret mount code
