+++ services/start-local-dev-server.js
@@ -1,12 +1,5 @@
 const { AzureCliCredential } = require('@azure/identity');
 const { SecretClient } = require('@azure/keyvault-secrets');
-const app = require('../app');
-const config = require('config');
-const https = require('https');
-const logger = require('../logger');
-const webpack = require('webpack');
-const webpackDevConfig = require('../webpack/webpack.dev');
-const webpackMiddleware = require('webpack-dev-middleware');
 
 async function mount(vaultName, secret) {
   try {
@@ -27,26 +20,10 @@ async function mount(vaultName, secret) {
   }
 }
 
-async function fetchSecrets() {
+async function run() {
   const serverKey = await mount('sscs-aat', 'server-key');
   const serverCertificate = await mount('sscs-aat', 'server-certificate');
   return { serverKey, serverCertificate };
 }
 
-const startLocalDevServer = async() => {
-  const compiler = webpack(webpackDevConfig);
-  const wp = webpackMiddleware(compiler, { publicPath: webpackDevConfig.output.publicPath });
-  app.use(wp);
-  const { serverKey, serverCertificate } = await fetchSecrets();
-  wp.waitUntilValid(stats => {
-    app.locals.webpackHash = stats.hash;
-    https.createServer({
-      key: serverKey,
-      cert: serverCertificate
-    }, app).listen(config.node.port, () => {
-      logger.trace(`SYA server listening on port: ${config.node.port}`);
-    });
-  });
-};
-
-module.exports = startLocalDevServer;
+module.exports = run;
